This is an initial submittal of a proposed new client API. It takes ideas from Jean-Francois, Simeon and myself. It is based upon work done in the original Wiseman client API, a low level client stub factory created by Jean-Francois, and the "Settings/BuildMessage" classes from Simeon.
It is also partly due to the request from users for a higher level client API.

This work is just a proposal and is being submitted into the client_api_proposal
branch in order to get feedback on its direction.

While working on this I have tried to keep the following goals in mind:

   * Make it easily extendable
   * Make it able to support many clients: J2SE, JAX-WS, & Apache HttpClient
   * Make it simpler & more intuitive to use
   * Retain the power to easily set all possible options in a request

At the lowest level, base upon work from Jean-Francois, I have:

   * WSManMessageFactory
   * SOAPRequest interface
   * SOAPResponse interface

There are impelmentations for J2SE & JAX-WS. Default is currently set to use
JAX-WS.

Next I have a middle layer set of message classes. I have created a class for every
operation, e.g. Create, Get, Update, Delete, Enumerate, Pull, Release, etc...

Some classes, such as those for Eventing, are just stubs and not yet complete.
This is more to give you an idea where this is headed. Look into the "com.sun.ws.management.client.message" package for details. Sorry, there is
little to no javadoc at the moment.

The base classes are WSAddressingRequest & WSAddressingResponse. They implement
the SOAPRequest & SOAPResponss interfaces and use the message factory
during construction. All the other message classes are built upon these
two classes. The hierarchy is as follows:

  - SOAPRequest interface
    - WSAddressingRequest
      - WSManRequest
        - WsManTransferRequest
          - WSManCreateRequest
          - WSManGetRequest
          - WSManPutRequest
          - WSManDeleteRequest
        - WSManEnumerationRequest
          - WSManEnumerateRequest
          - WSManPullRequest
          - WSManReleaseRequest
        - WSManEventingRequest
          ...

  - SOAPResponse interface
    - WSAddressingResponse
      - WSManResponse
        - WsManTransferResponse
          - WSManCreateResponse
          - WSManGetResponse
          - WSManPutResponse
          - WSManDeleteResponse
        - WSManEnumerationResponse
          - WSManEnumerateResponse
          - WSManPullResponse
          - WSManReleaseResponse
        - WSManEventingResponse
          ...


The classes, for the most part, do not depend upon the current classes such as
Management, Trasfer, TransferExtensions, Enumeration, EnumerationsExtensions, etc...
Rather, any code needed from them has been copied over to create an independent set
of client classes.

These classes are easily extendable, so someone could easily create a set of
custom classes, for eample the Traffic Light resource:

          - TrafficLightCreateRequest (subclass of WSManCreateRequest)
          - TrafficLightGetRequest (subclass of WSManGetRequest)
          - TrafficLightPutRequest (subclass of WSManPutRequest)
          - TrafficLightDeleteRequest (subclass of WSManDeleteRequest)
          - etc...

These could also be easily generated automatically for the user.

It also makes it easy to create classes for custom operations.


A JUnit test has been added to test Create/Get/Delete. That is the only parts
that have been tested, but I ran a test for both JAX-WS & J2SE. Both work.

At the higher layer is a modified version of an API I already had, but was not
part of Wiseman:

   * WSManClient
   * ResourceIterator

This has only partially benn ported to use the newer message classes, but I
think this can give you an idea of what the highest level will look like. The
ResourceIterator does not work at all at the moment, but I think the stubs
and the javadoc can give you an idea on how it would work. You just basically
create an iterator and it takes care of all the enumerate/pull/release messages
for you.

This is my last week before 3 weeks of vacation and I will not have a lot of
time to work on this before my vacation, so you can take you time on feedback.
I'll work on this again when I return in March.

Any and all feedback is welcome.

Denis Rachal